transactional #finish

+++ mysite03/src/main/java/mysite/service/GuestbookService.java
@@ -11,8 +11,8 @@ import mysite.vo.GuestbookVo;
 
 @Service
 public class GuestbookService {
-	private GuestbookRepository guestbookRepository;
-	private GuestbookLogRepository guestbookLogRepository;
+	private final GuestbookRepository guestbookRepository;
+	private final GuestbookLogRepository guestbookLogRepository;
 	
 	public GuestbookService(GuestbookRepository guestbookRepository, GuestbookLogRepository guestbookLogRepository) {
 		this.guestbookRepository = guestbookRepository;
@@ -25,11 +25,23 @@ public class GuestbookService {
 	
 	@Transactional
 	public void deleteContents(Long id, String password) {
-		guestbookRepository.deleteByIdAndPassword(id, password);
+		GuestbookVo vo = guestbookRepository.findById(id);
+		if(vo == null) {
+			return;
+		}
+		
+		int count = guestbookRepository.deleteByIdAndPassword(id, password);
+		if(count == 1) {
+			guestbookLogRepository.updateByRegDate(vo.getRegDate());
+		}
 	}
 	
 	@Transactional
 	public void addContents(GuestbookVo vo) {
+		int count = guestbookLogRepository.update();
+		if(count == 0) {
+			guestbookLogRepository.insert();
+		}
 		guestbookRepository.insert(vo);
 	}
 }
